requirements.txt
Django==1.10.4
django-tastypie==0.13.3
python-dateutil==2.6.0
python-mimeparse==1.6.0
six==1.10.0
